Simple tests of new interactive EBSD detector plots
