+++ tests/test_signals/test_ebsd_master_pattern.py
@@ -543,6 +543,9 @@ class TestProjectFromLambert:
 
 
 class TestMasterPatternPlotting:
+    @pytest.mark.skipif(
+        dependency_version["pyvista"] is None, reason="PyVista is not installed"
+    )
     def test_plot_spherical(self, skipif_no_vtk_support):
         """Returns expected data and raises correct error."""
         import pyvista as pv
